backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.3
pandas==2.2.1
numpy==1.26.4
yfinance==0.2.37
scikit-learn==1.4.1.post1
apscheduler==3.10.4
pytest==8.1.1
httpx==0.27.0
textblob==0.18.0
vaderSentiment==3.3.2
beautifulsoup4==4.12.3
requests==2.31.0
feedparser==6.0.11
python-dotenv==1.0.1
python-jose[cryptography]
python-multipart
